feat: first commit

+++ src/use-cases/delete-user.ts
@@ -1,4 +1,5 @@
 import { UserRepository } from "../repositories/user-repository";
+import { UserNotFound } from "./errors/user-not-foud";
 
 interface DeleteUserUseCaseRequest {
 	id: string;
@@ -10,6 +11,10 @@ export class DeleteUserUseCase {
 	constructor(private useRepository: UserRepository) {}
 
 	async execute({ id }: DeleteUserUseCaseRequest): Promise<DeleteUserUseCaseResponse> {
+		const user = await this.useRepository.findById(id);
+		if (!user) {
+			throw new UserNotFound();
+		}
 		return await this.useRepository.delete(id);
 	}
 }
